link to /graphs/contributors instead of requiring manual updates
In order to prevent https://github.com/jonschlinkert/maintainers-guide-to-staying-positive#contributors from becoming stale, perhaps a link to https://github.com/jonschlinkert/maintainers-guide-to-staying-positive/graphs/contributors is more appropriate. The "graphs" page provides the same information, and is guaranteed to be current without any manual intervention, even if it is not updated immediately. 

In my experience, anything like this that can be automated should be automated. Contributors can then focus more on contributing and less on repo-specific process.